+++ stream/benchmark_test.go
@@ -15,6 +15,6 @@ func BenchmarkSeeOtherHostError(b *testing.B) {
 	ip := &net.IPAddr{IP: net.ParseIP("2001:db8::68")}
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
-		_ = stream.SeeOtherHostError(ip, nil)
+		_ = stream.SeeOtherHostError(ip)
 	}
 }